Replace latch with cpp20_latch
