+++ libs/pika/executors/include/pika/executors/detail/hierarchical_spawning.hpp
@@ -60,7 +60,7 @@ namespace pika { namespace parallel { namespace execution { namespace detail {
         auto post_policy = policy;
         post_policy.set_stacksize(pika::execution::thread_stacksize::small_);
 
-        lcos::local::latch l(size);
+        pika::latch l(size);
         std::size_t part_begin = 0;
         auto it = std::begin(shape);
         for (std::size_t t = 0; t < num_threads; ++t)
