+++ libs/pika/executors/tests/unit/annotating_executor.cpp
@@ -25,7 +25,7 @@
 ///////////////////////////////////////////////////////////////////////////////
 std::string annotation;
 
-void test_post_f(int passed_through, pika::lcos::local::latch& l)
+void test_post_f(int passed_through, pika::latch& l)
 {
     PIKA_TEST_EQ(passed_through, 42);
 
@@ -44,7 +44,7 @@ void test_post(Executor&& executor)
         auto exec = pika::experimental::prefer(
             pika::execution::experimental::with_annotation, executor, desc);
 
-        pika::lcos::local::latch l(2);
+        pika::latch l(2);
         pika::parallel::execution::post(exec, &test_post_f, 42, std::ref(l));
         l.arrive_and_wait();
 
@@ -58,7 +58,7 @@ void test_post(Executor&& executor)
         auto exec =
             pika::execution::experimental::with_annotation(executor, desc);
 
-        pika::lcos::local::latch l(2);
+        pika::latch l(2);
         pika::parallel::execution::post(exec, &test_post_f, 42, std::ref(l));
         l.arrive_and_wait();
 
@@ -375,7 +375,7 @@ void test_post_policy(ExPolicy&& policy)
     std::string desc("test_post_policy");
     auto p = pika::execution::experimental::with_annotation(policy, desc);
 
-    pika::lcos::local::latch l(2);
+    pika::latch l(2);
     pika::parallel::execution::post(
         p.executor(), &test_post_f, 42, std::ref(l));
     l.arrive_and_wait();
