+++ libs/pika/executors/tests/unit/scheduler_executor.cpp
@@ -23,7 +23,7 @@
 ///////////////////////////////////////////////////////////////////////////////
 bool executed = false;
 
-void test_post_f(int passed_through, pika::lcos::local::latch& l)
+void test_post_f(int passed_through, pika::latch& l)
 {
     PIKA_TEST_EQ(passed_through, 42);
 
@@ -37,7 +37,7 @@ void test_post(Executor&& exec)
 {
     executed = false;
 
-    pika::lcos::local::latch l(2);
+    pika::latch l(2);
     pika::parallel::execution::post(exec, &test_post_f, 42, std::ref(l));
     l.arrive_and_wait();
