+++ libs/pika/synchronization/tests/unit/latch.cpp
@@ -1,13 +1,12 @@
-//  Copyright (c) 2015 Hartmut Kaiser
+//  Copyright (c) 2015-2020 Hartmut Kaiser
 //
 //  SPDX-License-Identifier: BSL-1.0
 //  Distributed under the Boost Software License, Version 1.0. (See accompanying
 //  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
-#include <pika/execution.hpp>
-#include <pika/future.hpp>
 #include <pika/init.hpp>
 #include <pika/latch.hpp>
+#include <pika/modules/async.hpp>
 #include <pika/testing.hpp>
 
 #include <atomic>
@@ -21,43 +20,44 @@
 std::atomic<std::size_t> num_threads(0);
 
 ///////////////////////////////////////////////////////////////////////////////
-void test_count_down_and_wait(pika::lcos::local::latch& l)
+void test_arrive_and_wait(pika::latch& l)
 {
     ++num_threads;
 
-    PIKA_TEST(!l.is_ready());
-    l.count_down_and_wait();
+    PIKA_TEST(!l.try_wait());
+    l.arrive_and_wait();
 }
 
-void test_count_down(pika::lcos::local::latch& l)
+void test_count_down(pika::latch& l)
 {
     ++num_threads;
 
-    PIKA_TEST(!l.is_ready());
+    PIKA_TEST(!l.try_wait());
     l.count_down(NUM_THREADS);
 }
 
 ///////////////////////////////////////////////////////////////////////////////
 int pika_main()
 {
-    // count_down_and_wait
+    // arraive_and_wait
     {
-        pika::lcos::local::latch l(NUM_THREADS + 1);
-        PIKA_TEST(!l.is_ready());
+        pika::latch l(NUM_THREADS + 1);
+        PIKA_TEST(!l.try_wait());
 
         std::vector<pika::future<void>> results;
         for (std::ptrdiff_t i = 0; i != NUM_THREADS; ++i)
-            results.push_back(
-                pika::async(&test_count_down_and_wait, std::ref(l)));
+        {
+            results.push_back(pika::async(&test_arrive_and_wait, std::ref(l)));
+        }
 
-        PIKA_TEST(!l.is_ready());
+        PIKA_TEST(!l.try_wait());
 
         // Wait for all threads to reach this point.
-        l.count_down_and_wait();
+        l.arrive_and_wait();
 
         pika::wait_all(results);
 
-        PIKA_TEST(l.is_ready());
+        PIKA_TEST(l.try_wait());
         PIKA_TEST_EQ(num_threads.load(), NUM_THREADS);
     }
 
@@ -65,17 +65,17 @@ int pika_main()
     {
         num_threads.store(0);
 
-        pika::lcos::local::latch l(NUM_THREADS + 1);
-        PIKA_TEST(!l.is_ready());
+        pika::latch l(NUM_THREADS + 1);
+        PIKA_TEST(!l.try_wait());
 
         pika::future<void> f = pika::async(&test_count_down, std::ref(l));
 
-        PIKA_TEST(!l.is_ready());
-        l.count_down_and_wait();
+        PIKA_TEST(!l.try_wait());
+        l.arrive_and_wait();
 
         f.get();
 
-        PIKA_TEST(l.is_ready());
+        PIKA_TEST(l.try_wait());
         PIKA_TEST_EQ(num_threads.load(), std::size_t(1));
     }
 
@@ -83,71 +83,20 @@ int pika_main()
     {
         num_threads.store(0);
 
-        pika::lcos::local::latch l(NUM_THREADS);
-        PIKA_TEST(!l.is_ready());
+        pika::latch l(NUM_THREADS);
+        PIKA_TEST(!l.try_wait());
 
         std::vector<pika::future<void>> results;
         for (std::ptrdiff_t i = 0; i != NUM_THREADS; ++i)
-            results.push_back(
-                pika::async(&test_count_down_and_wait, std::ref(l)));
+        {
+            results.push_back(pika::async(&test_arrive_and_wait, std::ref(l)));
+        }
 
         pika::wait_all(results);
 
         l.wait();
 
-        PIKA_TEST(l.is_ready());
-        PIKA_TEST_EQ(num_threads.load(), NUM_THREADS);
-    }
-
-    // count_up
-    {
-        num_threads.store(0);
-
-        pika::lcos::local::latch l(1);
-        PIKA_TEST(!l.is_ready());
-
-        std::vector<pika::future<void>> results;
-        for (std::ptrdiff_t i = 0; i != NUM_THREADS; ++i)
-        {
-            l.count_up(1);
-            results.push_back(
-                pika::async(&test_count_down_and_wait, std::ref(l)));
-        }
-
-        PIKA_TEST(!l.is_ready());
-
-        // Wait for all threads to reach this point.
-        l.count_down_and_wait();
-
-        pika::wait_all(results);
-
-        PIKA_TEST(l.is_ready());
-        PIKA_TEST_EQ(num_threads.load(), NUM_THREADS);
-    }
-
-    // reset
-    {
-        num_threads.store(0);
-
-        pika::lcos::local::latch l(0);
-        l.reset(NUM_THREADS + 1);
-        PIKA_TEST(!l.is_ready());
-
-        std::vector<pika::future<void>> results;
-        for (std::ptrdiff_t i = 0; i != NUM_THREADS; ++i)
-        {
-            results.push_back(
-                pika::async(&test_count_down_and_wait, std::ref(l)));
-        }
-
-        PIKA_TEST(!l.is_ready());
-
-        // Wait for all threads to reach this point.
-        l.count_down_and_wait();
-
-        pika::wait_all(results);
-
-        PIKA_TEST(l.is_ready());
+        PIKA_TEST(l.try_wait());
         PIKA_TEST_EQ(num_threads.load(), NUM_THREADS);
     }
